pyproject.toml
[project]
name = "jellyseek"
version = "0.0.1"
description = "Deepseek RAG implamentation of jellyfin searches"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.0",
    "python-dotenv>=1.0.1"
]
